% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basic_variant_calc}
\alias{basic_variant_calc}
\title{Basic variant calculate}
\usage{
basic_variant_calc(data, pcnt, u_value)
}
\arguments{
\item{data}{\itemize{
\item dataframe used in the calculation
}}

\item{pcnt}{\itemize{
\item the percentage parameter
}}

\item{u_value}{\itemize{
\item the user-defined parameter
}}
}
\value{
data - dataframe with the basic_variant calc
}
\description{
This function calculates the percentage of the original value
}
