% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate}
\alias{calculate}
\title{Calculate results}
\usage{
calculate(data)
}
\arguments{
\item{data}{Data frame used for calculation. Needs to contain 5 columns named a-e (see \code{\link{data}} data set for more info)}
}
\value{
Input data frame with additional column \code{res} that contains calculation results for rows
}
\description{
Calculate results using given data on a row basis
}
