% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example data set for Senior Software Developer recruitment task}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{a}{variable a, range 10000-50000}
\item{b}{variable b, range 50-100}
\item{c}{variable c, range 10-20}
\item{d}{variable d, range 1-5}
\item{e}{variable e, range 1-2}
}
}
\usage{
data
}
\description{
This data set simulates data that could be provided for some analytics work.
You can assume that each row represents some data entry (e.g. single day),
and each column is some variable that affects calculations of results
(see \code{\link{calculate}} function for more info).
}
\keyword{datasets}
