% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{advanced_variant_calc}
\alias{advanced_variant_calc}
\title{Advanced variant calculate}
\usage{
advanced_variant_calc(data, u_value)
}
\arguments{
\item{data}{\itemize{
\item dataframe used in the calculation
}}

\item{u_value}{\itemize{
\item the user-defined parameter
}}
}
\value{
data - dataframe with the advanced variant calc
}
\description{
This function the original value with user-defined value
}
